requirement.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
decorator==5.1.0
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
future==0.18.2
huggingface-hub==0.0.16
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
librosa==0.8.1
llvmlite==0.37.0
MarkupSafe==2.0.1
numba==0.54.0
numpy==1.20.3
packaging==21.0
pooch==1.5.1
pycparser==2.20
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
resampy==0.2.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
tqdm==4.62.2
transformers==4.10.0
typing-extensions==3.10.0.2
urllib3==1.26.6
waitress==2.0.0
Werkzeug==2.0.1
